Remove Developer Certificate of Origin
The Developer Certificate of Origin (DCO) and "whole-project license"
were removed as they were determined to be of little benefit. This
change brings the contribution requirements more in line with the
OpenBSD project's.

The main difference is that we will not be accepting contributions not
licensed under the verbatim OpenBSD/ISC license indicated in the Esirra
README file.